Reorg of pipeline
