+++ morpheus/pipeline/pipeline.py
@@ -33,811 +33,17 @@ import cudf
 from morpheus.config import Config
 from morpheus.config import CppConfig
 from morpheus.messages import MultiMessage
+from morpheus.pipeline.receiver import Receiver
+from morpheus.pipeline.sender import Sender
+from morpheus.pipeline.source_stage import SourceStage
+from morpheus.pipeline.stage import Stage
+from morpheus.pipeline.stream_wrapper import StreamWrapper
 from morpheus.utils.atomic_integer import AtomicInteger
 from morpheus.utils.type_utils import _DecoratorType
 from morpheus.utils.type_utils import greatest_ancestor
 from morpheus.utils.type_utils import pretty_print_type_name
 
-config = Config()
-
 logger = logging.getLogger(__name__)
-
-
-def get_time_ms():
-    return round(time.time() * 1000)
-
-
-T = typing.TypeVar('T')
-
-StreamPair = typing.Tuple[neo.Node, typing.Type]
-
-
-class Sender():
-    """
-    The `Sender` object represents a port on a `StreamWrapper` object that sends messages to a `Receiver`.
-
-    Parameters
-        ----------
-        parent : `morpheus.pipeline.pipeline.StreamWrapper`
-            Parent `StreamWrapper` object.
-        port_number : int
-            Sender port number.
-    """
-
-    def __init__(self, parent: "StreamWrapper", port_number: int):
-
-        self._parent = parent
-        self.port_number = port_number
-
-        self._output_receivers: typing.List[Receiver] = []
-
-        self._out_stream_pair: StreamPair = (None, None)
-
-    @property
-    def parent(self):
-        return self._parent
-
-    @property
-    def is_complete(self):
-        # Sender is complete when the type or stream has been set
-        return self._out_stream_pair != (None, None)
-
-    @property
-    def out_pair(self):
-        return self._out_stream_pair
-
-    @property
-    def out_stream(self):
-        return self._out_stream_pair[0]
-
-    @property
-    def out_type(self):
-        return self._out_stream_pair[1]
-
-
-class Receiver():
-    """
-    The `Receiver` object represents a downstream port on a `StreamWrapper` object that gets messages from a `Sender`.
-
-    Parameters
-        ----------
-        parent : `morpheus.pipeline.pipeline.StreamWrapper`
-            Parent `StreamWrapper` object.
-        port_number : int
-            Receiver port number.
-    """
-
-    def __init__(self, parent: "StreamWrapper", port_number: int):
-
-        self._parent = parent
-        self.port_number = port_number
-
-        self._is_linked = False
-
-        self._input_type = None
-        self._input_stream = None
-
-        self._input_senders: typing.List[Sender] = []
-
-    @property
-    def parent(self):
-        return self._parent
-
-    @property
-    def is_complete(self):
-        """
-        A receiver is complete if all input senders are also complete.
-        """
-        return all([x.is_complete for x in self._input_senders])
-
-    @property
-    def is_partial(self):
-        """
-        A receiver is partially complete if any input sender is complete. Receivers are usually partially complete if
-        there is a circular pipeline.
-        """
-        # Its partially complete if any input sender is complete
-        return any([x.is_complete for x in self._input_senders])
-
-    @property
-    def in_pair(self):
-        return (self.in_stream, self.in_pair)
-
-    @property
-    def in_stream(self):
-        return self._input_stream
-
-    @property
-    def in_type(self):
-        return self._input_type
-
-    def get_input_pair(self) -> StreamPair:
-
-        assert self.is_partial, "Must be partially complete to get the input pair!"
-
-        # Build the input from the senders
-        if (self._input_stream is None and self._input_type is None):
-            # First check if we only have 1 input sender
-            if (len(self._input_senders) == 1):
-                # In this case, our input stream/type is determined from the sole Sender
-                sender = self._input_senders[0]
-
-                self._input_stream = sender.out_stream
-                self._input_type = sender.out_type
-                self._is_linked = True
-            else:
-                # We have multiple senders. Create a dummy stream to connect all senders
-                if (self.is_complete):
-                    # Connect all streams now
-                    # self._input_stream = streamz.Stream(upstreams=[x.out_stream for x in self._input_senders],
-                    #                                     asynchronous=True,
-                    #                                     loop=IOLoop.current())
-                    raise NotImplementedError("Still using streamz")
-                    self._is_linked = True
-                else:
-                    # Create a dummy stream that needs to be linked later
-                    # self._input_stream = streamz.Stream(asynchronous=True, loop=IOLoop.current())
-                    raise NotImplementedError("Still using streamz")
-
-                # Now determine the output type from what we have
-                great_ancestor = greatest_ancestor(*[x.out_type for x in self._input_senders if x.is_complete])
-
-                if (great_ancestor is None):
-                    # TODO: Add stage, port, and type info to message
-                    raise RuntimeError(("Cannot determine single type for senders of input port. "
-                                        "Use a merge stage to handle different types of inputs."))
-
-                self._input_type = great_ancestor
-
-        return (self._input_stream, self._input_type)
-
-    def link(self):
-        """
-        The linking phase determines the final type of the `Receiver` and connects all underlying stages.
-
-        Raises:
-            RuntimeError: Throws a `RuntimeError` if the predicted input port type determined during the build phase is
-            different than the current port type.
-        """
-
-        assert self.is_complete, "Must be complete before linking!"
-
-        if (self._is_linked):
-            return
-
-        # Check that the types still work
-        great_ancestor = greatest_ancestor(*[x.out_type for x in self._input_senders if x.is_complete])
-
-        if (not typing_utils.issubtype(great_ancestor, self._input_type)):
-            # TODO: Add stage, port, and type info to message
-            raise RuntimeError(
-                "Invalid linking phase. Input port type does not match predicted type determined during build phase")
-
-        for out_stream in [x.out_stream for x in self._input_senders]:
-            out_stream.connect(self._input_stream)
-
-        self._is_linked = True
-
-
-def _save_init_vals(func: _DecoratorType) -> _DecoratorType:
-
-    # Save the signature only once
-    sig = inspect.signature(func, follow_wrapped=True)
-
-    def inner(self: "StreamWrapper", c: Config, *args, **kwargs):
-
-        # Actually call init first. This way any super classes strings will be overridden
-        func(self, c, *args, **kwargs)
-
-        # Determine all set values
-        bound = sig.bind(self, c, *args, **kwargs)
-        bound.apply_defaults()
-
-        init_pairs = []
-
-        for key, val in bound.arguments.items():
-
-            # We really dont care about these
-            if (key == "self" or key == "c"):
-                continue
-
-            init_pairs.append(f"{key}={val}")
-
-        # Save values on self
-        self._init_str = ", ".join(init_pairs)
-
-        return
-
-    return typing.cast(_DecoratorType, inner)
-
-
-class StreamWrapper(ABC, collections.abc.Hashable):
-    """
-    This abstract class serves as the morpheus.pipeline's base class. This class wraps a `neo.Node`
-    object and aids in hooking stages up together.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-
-    """
-
-    __ID_COUNTER = AtomicInteger(0)
-
-    def __init__(self, c: Config):
-        self._id = StreamWrapper.__ID_COUNTER.get_and_inc()
-        self._pipeline: Pipeline = None
-        self._init_str: str = ""  # Stores the initialization parameters used for creation. Needed for __repr__
-
-        # Indicates whether or not this wrapper has been built. Can only be built once
-        self._is_built = False
-
-        # Input/Output ports used for connecting stages
-        self._input_ports: typing.List[Receiver] = []
-        self._output_ports: typing.List[Sender] = []
-
-    def __init_subclass__(cls) -> None:
-
-        # Wrap __init__ to save the arg values
-        cls.__init__ = _save_init_vals(cls.__init__)
-
-        return super().__init_subclass__()
-
-    def __hash__(self) -> int:
-        return self._id
-
-    def __str__(self):
-        text = f"<{self.unique_name}; {self.__class__.__name__}({self._init_str})>"
-
-        return text
-
-    __repr__ = __str__
-
-    @property
-    @abstractmethod
-    def name(self) -> str:
-        """
-        The name of the stage. Used in logging. Each derived class should override this property with a unique
-        name.
-
-        Returns
-        -------
-        str
-            Name of a stage.
-
-        """
-        pass
-
-    @property
-    def unique_name(self) -> str:
-        """
-        Unique name of stag. Generated by appending stage id to stage name.
-
-        Returns
-        -------
-        str
-            Unique name of stage.
-        """
-        return f"{self.name}-{self._id}"
-
-    @property
-    def is_built(self) -> bool:
-        """
-        Indicates if this stage has been built.
-
-        Returns
-        -------
-        bool
-            True if stage is built, False otherwise.
-        """
-        return self._is_built
-
-    @property
-    def input_ports(self) -> typing.List[Receiver]:
-        """Input ports to this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.Receiver`]
-            Input ports to this stage.
-        """
-        return self._input_ports
-
-    @property
-    def output_ports(self) -> typing.List[Sender]:
-        """
-        Output ports from this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.Sender`]
-            Output ports from this stage.
-        """
-        return self._output_ports
-
-    @property
-    def has_multi_input_ports(self) -> bool:
-        """
-        Indicates if this stage has multiple input ports.
-
-        Returns
-        -------
-        bool
-            True if stage has multiple input ports, False otherwise.
-        """
-        return len(self._input_ports) > 1
-
-    @property
-    def has_multi_output_ports(self) -> bool:
-        """
-        Indicates if this stage has multiple output ports.
-
-        Returns
-        -------
-        bool
-            True if stage has multiple output ports, False otherwise.
-        """
-        return len(self._output_ports) > 1
-
-    def get_all_inputs(self) -> typing.List[Sender]:
-        """
-        Get all input senders to this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.Sender`]
-            All input senders.
-        """
-
-        senders = []
-
-        for in_port in self._input_ports:
-            senders.extend(in_port._input_senders)
-
-        return senders
-
-    def get_all_input_stages(self) -> typing.List["StreamWrapper"]:
-        """
-        Get all input stages to this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.StreamWrapper`]
-            All input stages.
-        """
-        return [x.parent for x in self.get_all_inputs()]
-
-    def get_all_outputs(self) -> typing.List[Receiver]:
-        """
-        Get all output receivers from this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.Receiver`]
-            All output receivers.
-        """
-        receivers = []
-
-        for out_port in self._output_ports:
-            receivers.extend(out_port._output_receivers)
-
-        return receivers
-
-    def get_all_output_stages(self) -> typing.List["StreamWrapper"]:
-        """
-        Get all output stages from this stage.
-
-        Returns
-        -------
-        typing.List[`morpheus.pipeline.pipeline.StreamWrapper`]
-            All output stages.
-        """
-        return [x.parent for x in self.get_all_outputs()]
-
-    def supports_cpp_node(self):
-        """
-        Specifies whether this Stage is even capable of creating C++ nodes. During the build phase, this value will be
-        combined with Config.get().use_cpp to determine whether or not a C++ node is created. This is an instance method
-        to allow runtime decisions and derived classes to override base implementations.
-        """
-        # By default, return False unless otherwise specified
-        return False
-
-    def _build_cpp_node(self):
-        """
-        Specifies whether or not to build a C++ node. Only should be called during the build phase.
-        """
-        return CppConfig.get_should_use_cpp() and self.supports_cpp_node()
-
-    def can_build(self, check_ports=False) -> bool:
-        """
-        Determines if all inputs have been built allowing this node to be built.
-
-        Parameters
-        ----------
-        check_ports : bool, optional
-            Check if we can build based on the input ports, by default False.
-
-        Returns
-        -------
-        bool
-            True if we can build, False otherwise.
-        """
-
-        # Can only build once
-        if (self.is_built):
-            return False
-
-        if (not check_ports):
-            # We can build if all input stages have been built. Easy and quick check. Works for non-circular pipelines
-            for in_stage in self.get_all_input_stages():
-                if (not in_stage.is_built):
-                    return False
-
-            return True
-        else:
-            # Check if we can build based on the input ports. We can build
-            for r in self.input_ports:
-                if (not r.is_partial):
-                    return False
-
-            return True
-
-    def build(self, seg: neo.Segment, do_propagate=True):
-        """Build this stage.
-
-        Parameters
-        ----------
-        seg : `neo.Segment`
-            Neo segment for this stage.
-        do_propagate : bool, optional
-            Whether to propagate to build output stages, by default True.
-
-        """
-        assert not self.is_built, "Can only build stages once!"
-        assert self._pipeline is not None, "Must be attached to a pipeline before building!"
-
-        # Pre-Build returns the input pairs for each port
-        in_ports_pairs = self._pre_build()
-
-        out_ports_pair = self._build(seg, in_ports_pairs)
-
-        # Allow stages to do any post build steps (i.e., for sinks, or timing functions)
-        out_ports_pair = self._post_build(seg, out_ports_pair)
-
-        assert len(out_ports_pair) == len(self.output_ports), \
-            "Build must return same number of output pairs as output ports"
-
-        # Assign the output ports
-        for port_idx, out_pair in enumerate(out_ports_pair):
-            self.output_ports[port_idx]._out_stream_pair = out_pair
-
-        self._is_built = True
-
-        if (not do_propagate):
-            return
-
-        # Now build for any dependents
-        for dep in self.get_all_output_stages():
-            if (not dep.can_build()):
-                continue
-
-            dep.build(seg, do_propagate=do_propagate)
-
-    def _pre_build(self) -> typing.List[StreamPair]:
-        in_pairs: typing.List[StreamPair] = [x.get_input_pair() for x in self.input_ports]
-
-        return in_pairs
-
-    @abstractmethod
-    def _build(self, seg: neo.Segment, in_ports_streams: typing.List[StreamPair]) -> typing.List[StreamPair]:
-        """
-        This function is responsible for constructing this stage's internal `neo.Node` object. The input
-        of this function contains the returned value from the upstream stage.
-
-        The input values are the `neo.Segment` for this stage and a `StreamPair` tuple which contain the input
-        `neo.Node` object and the message data type.
-
-        :meta public:
-
-        Parameters
-        ----------
-        seg : `neo.Segment`
-            `neo.Segment` object for the pipeline. This should be used to construct/attach the internal `neo.Node`.
-        in_ports_streams : `morpheus.pipeline.pipeline.StreamPair`
-            List of tuples containing the input `neo.Node` object and the message data type.
-
-        Returns
-        -------
-        `typing.List[morpheus.pipeline.pipeline.StreamPair]`
-            List of tuples containing the output `neo.Node` object from this stage and the message data type.
-
-        """
-        pass
-
-    def _post_build(self, seg: neo.Segment, out_ports_pair: typing.List[StreamPair]) -> typing.List[StreamPair]:
-        return out_ports_pair
-
-    def start(self):
-
-        assert self.is_built, "Must build before starting!"
-
-        self._start()
-
-    def _start(self):
-        pass
-
-    def stop(self):
-        """
-        Stages can implement this to perform cleanup steps when pipeline is stopped.
-        """
-        pass
-
-    async def join(self):
-        pass
-
-    def _create_ports(self, input_count: int, output_count: int):
-        assert len(self._input_ports) == 0 and len(self._output_ports) == 0, "Can only create ports once!"
-
-        self._input_ports = [Receiver(parent=self, port_number=i) for i in range(input_count)]
-        self._output_ports = [Sender(parent=self, port_number=i) for i in range(output_count)]
-
-
-class SourceStage(StreamWrapper):
-    """
-    The SourceStage is mandatory for the Morpheus pipeline to run. This stage represents the start of the pipeline. All
-    `SourceStage` object take no input but generate output.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-
-    """
-
-    def __init__(self, c: Config):
-        super().__init__(c)
-
-        self._start_callbacks: typing.List[typing.Callable] = []
-        self._stop_callbacks: typing.List[typing.Callable] = []
-
-        self._source_stream: neo.Node = None
-
-    @property
-    def input_count(self) -> int:
-        """
-        Return None for no max intput count.
-
-        Returns
-        -------
-        int
-            Input count.
-
-        """
-        return None
-
-    @abstractmethod
-    def _build_source(self, seg: neo.Segment) -> StreamPair:
-        """
-        Abstract method all derived Source classes should implement. Returns the same value as `build`.
-
-        :meta public:
-
-        Returns
-        -------
-
-        `morpheus.pipeline.pipeline.StreamPair`:
-            A tuple containing the output `neo.Node` object from this stage and the message data type.
-        """
-
-        pass
-
-    @typing.final
-    def _build(self, seg: neo.Segment, in_ports_streams: typing.List[StreamPair]) -> typing.List[StreamPair]:
-        # Derived source stages should override `_build_source` instead of this method. This allows for tracking the
-        # True source object separate from the output stream. If any other operators need to be added after the source,
-        # use `_post_build`
-        assert len(self.input_ports) == 0, "Sources shouldnt have input ports"
-
-        source_pair = self._build_source(seg)
-
-        curr_source = source_pair[0]
-
-        self._source_stream = curr_source
-
-        # Now setup the output ports
-        self._output_ports[0]._out_stream_pair = source_pair
-
-        return [source_pair]
-
-    def _post_build(self, seg: neo.Segment, out_ports_pair: typing.List[StreamPair]) -> typing.List[StreamPair]:
-
-        return out_ports_pair
-
-    def _start(self):
-        self._source_stream.start()
-
-    def stop(self):
-        self._source_stream.stop()
-
-    async def join(self):
-        self._source_stream.join()
-
-
-class SingleOutputSource(SourceStage):
-    """
-    Subclass of SourceStage for building source stages that generate output for single port.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-    """
-
-    def __init__(self, c: Config):
-        super().__init__(c)
-
-        self._create_ports(0, 1)
-
-    def _post_build_single(self, seg: neo.Segment, out_pair: StreamPair) -> StreamPair:
-        return out_pair
-
-    @typing.final
-    def _post_build(self, seg: neo.Segment, out_ports_pair: typing.List[StreamPair]) -> typing.List[StreamPair]:
-
-        ret_val = self._post_build_single(seg, out_ports_pair[0])
-
-        logger.info("Added source: {}\n  └─> {}".format(str(self), pretty_print_type_name(ret_val[1])))
-
-        return [ret_val]
-
-
-class Stage(StreamWrapper):
-    """
-    This class serves as the base for all pipeline stage implementations that are not source objects.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-
-    """
-
-    def __init__(self, c: Config):
-        super().__init__(c)
-
-    def _post_build(self, seg: neo.Segment, out_ports_pair: typing.List[StreamPair]) -> typing.List[StreamPair]:
-
-        return out_ports_pair
-
-    def _start(self):
-        pass
-
-    def on_start(self):
-        """
-        This function can be overridden to add usecase-specific implementation at the start of any stage in
-        the pipeline.
-        """
-        pass
-
-    def _on_complete(self, stream):
-
-        logger.info("Stage Complete: {}".format(self.name))
-
-
-class SinglePortStage(Stage):
-    """
-    Class used for building stages with single input port and single output port.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-    """
-
-    def __init__(self, c: Config):
-        super().__init__(c)
-
-        self._create_ports(1, 1)
-
-    @abstractmethod
-    def accepted_types(self) -> typing.Tuple:
-        """
-        Accepted input types for this stage are returned. Derived classes should override this method. An
-        error will be generated if the input types to the stage do not match one of the available types
-        returned from this method.
-
-        Returns
-        -------
-        typing.Tuple
-            Accepted input types.
-
-        """
-        pass
-
-    def _pre_build(self) -> typing.List[StreamPair]:
-        in_ports_pairs = super()._pre_build()
-
-        # Check the types of all inputs
-        for x in in_ports_pairs:
-            if (not typing_utils.issubtype(x[1], typing.Union[self.accepted_types()])):
-                raise RuntimeError("The {} stage cannot handle input of {}. Accepted input types: {}".format(
-                    self.name, x[1], self.accepted_types()))
-
-        return in_ports_pairs
-
-    @abstractmethod
-    def _build_single(self, seg: neo.Segment, input_stream: StreamPair) -> StreamPair:
-        pass
-
-    def _build(self, seg: neo.Segment, in_ports_streams: typing.List[StreamPair]) -> typing.List[StreamPair]:
-        # Derived source stages should override `_build_source` instead of this method. This allows for tracking the
-        # True source object separate from the output stream. If any other operators need to be added after the source,
-        # use `_post_build`
-        assert len(self.input_ports) == 1 and len(self.output_ports) == 1, \
-            "SinglePortStage must have 1 input port and 1 output port"
-
-        assert len(in_ports_streams) == 1, "Should only have 1 port on input"
-
-        return [self._build_single(seg, in_ports_streams[0])]
-
-    def _post_build_single(self, seg: neo.Segment, out_pair: StreamPair) -> StreamPair:
-        return out_pair
-
-    @typing.final
-    def _post_build(self, seg: neo.Segment, out_ports_pair: typing.List[StreamPair]) -> typing.List[StreamPair]:
-
-        ret_val = self._post_build_single(seg, out_ports_pair[0])
-
-        logger.info("Added stage: {}\n  └─ {} -> {}".format(str(self),
-                                                            pretty_print_type_name(self.input_ports[0].in_type),
-                                                            pretty_print_type_name(ret_val[1])))
-
-        return [ret_val]
-
-
-class MultiMessageStage(SinglePortStage):
-    """
-    Subclass of SinglePortStage with option to log timestamps in MessageMeta dataframe.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-    """
-
-    def __init__(self, c: Config):
-
-        # Derived classes should set this to true to log timestamps in debug builds
-        self._should_log_timestamps = False
-
-        super().__init__(c)
-
-    def _post_build_single(self, seg: neo.Segment, out_pair: StreamPair) -> StreamPair:
-
-        # Check if we are debug and should log timestamps
-        if (config.debug and self._should_log_timestamps):
-            # Cache the name property. Removes dependency on self in callback
-            cached_name = self.name
-
-            logger.info("Adding timestamp info for stage: '%s'", cached_name)
-
-            def post_timestamps(x: MultiMessage):
-
-                curr_time = get_time_ms()
-
-                x.set_meta("_ts_" + cached_name, curr_time)
-
-            # Only have one port
-            post_ts = seg.make_node(self.unique_name + "-ts", post_timestamps)
-            seg.make_edge(out_pair[0], post_ts)
-
-            # Keep the type unchanged
-            out_pair = (post_ts, out_pair[1])
-
-        return super()._post_build_single(seg, out_pair)
-
 
 class Pipeline():
     """
@@ -1237,71 +443,3 @@ class Pipeline():
         # Use asyncio.run() to launch the pipeline. This creates and destroys an event loop so re-running a pipeline in
         # the same process wont fail
         asyncio.run(self._do_run())
-
-
-class LinearPipeline(Pipeline):
-    """
-    This class is used to build linear pipelines where we have a single output source stage followed by stages that are
-    executed sequentially in the order they were added.
-
-    Parameters
-    ----------
-    c : `morpheus.config.Config`
-        Pipeline configuration instance.
-    """
-
-    def __init__(self, c: Config):
-        super().__init__(c)
-
-        self._linear_stages: typing.List[StreamWrapper] = []
-
-    def set_source(self, source: SourceStage):
-        """
-        Set a pipeline's source stage to consume messages before it begins executing stages. This must be
-        called once before `build_and_start`.
-
-        Parameters
-        ----------
-        source : `SourceStage`
-            The source stage wraps the implementation in a stream that allows it to read from Kafka or a file.
-
-        """
-
-        if (len(self._sources) > 0 and source not in self._sources):
-            logger.warning(
-                "LinearPipeline already has a source. Setting a new source will clear out all existing stages")
-
-            self._sources.clear()
-
-        # Store the source in sources
-        self._sources.add(source)
-
-        if (len(self._linear_stages) > 0):
-            logger.warning("Clearing %d stages from pipeline", len(self._linear_stages))
-            self._linear_stages.clear()
-
-        # Need to store the source in the pipeline
-        super().add_node(source)
-
-        # Store this as the first one in the linear stages. Must be index 0
-        self._linear_stages.append(source)
-
-    def add_stage(self, stage: SinglePortStage):
-        """
-        Add stages to the pipeline. All `Stage` classes added with this method will be executed sequentially
-        inthe order they were added.
-
-        Parameters
-        ----------
-        stage : `Stage`
-            The stage object to add. It cannot be already added to another `Pipeline` object.
-
-        """
-
-        assert len(self._linear_stages) > 0, "A source must be set on a LinearPipeline before adding any stages"
-        assert isinstance(stage, SinglePortStage), "Only `SinglePortStage` stages are accepted in `add_stage()`"
-
-        # Make an edge between the last node and this one
-        self.add_edge(self._linear_stages[-1], stage)
-
-        self._linear_stages.append(stage)
