+++ morpheus/pipeline/stream_pair.py
@@ -1,0 +1,6 @@
+
+import typing
+
+import neo
+
+StreamPair = typing.Tuple[neo.Node, typing.Type]
